Remove Non Usefule Code
